+++ Assets/Neoxider/Scripts/Cards/Model/BoardModel.cs
@@ -18,3 +18,4 @@ namespace Neo.Cards
         }
     }
 }
+