+++ Assets/Neoxider/Scripts/NPC/Navigation/NpcAggroFollowCore.cs
@@ -67,3 +67,4 @@ namespace Neo.NPC.Navigation
         }
     }
 }
+