+++ Assets/Neoxider/Scripts/Save/Providers/PlayerPrefsSaveProvider.cs
@@ -161,3 +161,4 @@ namespace Neo.Save
         }
     }
 }
+